refactored Item DB Calls, Availability Calls, Reservation Calls, User Calls
